Reimplemented directional movement with the sprites. A huge bug was with the BoundingBox struct which resulted in unwanted behavior. The summarize the bug, when initializing the bounding box, there was no need to sort such coordinates, applying if statements within a for loop to determine which digits where in the extreme sides was enough. Because I called a sort function on the coordinates, the actual far_top and far_left coordinates weren't the desired coordinates; essentially lied. The bug resulted in sprites leaving a trail of pixel characters instead of having natural movement with updating positions. For the movement, I replaced the swap() function with the two steps, reset(), which reseted the current sprite position, and then I call spawn() to draw the new position. I got rid of unnecessary code, organized code into their separate files. Formatted. Added some code comments. Instead of the game play, I added little demos to demonstrate the new updates. THE GAMEPLAY IS CURRENTLY UNDER CONSTRUCTION.

+++ src/entities/aliens.rs
@@ -24,7 +24,7 @@ pub struct Aliens {
 }
 
 impl Aliens {
-    pub fn new(engine: Rc<RefCell<Engine>>, count: usize, velocity: usize) -> Result<Self, Error> {
+    pub fn new(engine: Rc<RefCell<Engine>>, count: usize, velocity: f64) -> Result<Self, Error> {
         if count == 0 {
             return Err(Error::new(ErrorKind::Other, "Alien count cannot be 0"));
         }
